+++ test/model_test.jl
@@ -42,7 +42,8 @@ using RELOG, Cbc, JuMP, Printf, JSON, MathOptInterface.FileFormats
     end
 
     @testset "solve" begin
-        solution = RELOG.solve("$(pwd())/../instances/s1.json")
+        solution = RELOG.solve("$(pwd())/../instances/s1.json",
+                               output_filename="$(pwd())/../tmp/sol.json")
         
         @test "Costs" in keys(solution)
         @test "Fixed operating (\$)" in keys(solution["Costs"])