+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 0.3.3 (Aug 13, 2020)
+
+- Add option to write solution to JSON file in RELOG.solve
+- Improve error message when instance is infeasible
+- Make output file more readable
+
 # Version 0.3.2 (Aug 7, 2020)
 
 - Add "building period" parameter